refactor tokenpicker
