+++ src/Components/ActionsModal/components/ApproveSpendTokens/ApproveSpendTokens.tsx
@@ -6,7 +6,7 @@ import { useTranslation } from 'react-i18next';
 
 import Input from 'old-components/Guilds/common/Form/Input';
 import Avatar from 'old-components/Guilds/Avatar';
-import TokenPicker from 'old-components/Guilds/TokenPicker';
+import { TokenPicker } from 'Components';
 import TokenAmountInput from 'old-components/Guilds/common/Form/TokenAmountInput';
 import { useERC20Info } from 'hooks/Guilds/ether-swr/erc20/useERC20Info';
 import { useTokenList } from 'hooks/Guilds/tokens/useTokenList';
