+++ src/Components/TokenPicker/TokenPicker.styled.ts
@@ -1,0 +1,15 @@
+import styled from 'styled-components';
+
+export const TokenPickerContainer = styled.div`
+  margin: 2rem;
+`;
+
+export const SearchWrapper = styled.div`
+  display: flex;
+  margin-bottom: 1.5rem;
+`;
+
+export const TokenList = styled.div`
+  display: flex;
+  flex-direction: column;
+`;
