+++ src/Components/TokenPicker/fixtures.ts
@@ -1,0 +1,41 @@
+export const data = [
+  {
+    address: '0xB58b358Be2B16Cb8992fc7340F92127083de33cE',
+    name: 'SWAPR Guild',
+    decimals: 18,
+    symbol: 'SWPR',
+    fetchPrice: true,
+    logoURI: 'https://s2.coinmarketcap.com/static/img/coins/200x200/5589.png',
+    chainId: 1337,
+    balance: {
+      type: 'BigNumber',
+      hex: '0x8ac7230489e80000',
+    },
+  },
+  {
+    address: '0xfFb1cd0F95368DDd06D556161c5D3d9f0f4Fe6d2',
+    name: 'DXdao on Localhost',
+    decimals: 18,
+    symbol: 'DXD',
+    fetchPrice: true,
+    logoURI: 'https://s2.coinmarketcap.com/static/img/coins/200x200/5589.png',
+    chainId: 1337,
+    balance: {
+      type: 'BigNumber',
+      hex: '0x00',
+    },
+  },
+  {
+    address: '0x7BC0DEDaFd60611D430c89353CC30e1a11b90AC7',
+    name: 'REP Guild Token on Localhost',
+    decimals: 18,
+    symbol: 'RGT',
+    fetchPrice: true,
+    logoURI: 'https://s2.coinmarketcap.com/static/img/coins/200x200/5589.png',
+    chainId: 1337,
+    balance: {
+      type: 'BigNumber',
+      hex: '0x00',
+    },
+  },
+];
