+++ src/Components/TokenPicker/index.ts
@@ -1,0 +1,1 @@
+export { default as TokenPicker } from './TokenPicker';
