+++ src/Components/index.ts
@@ -1,3 +1,4 @@
 export { Header } from './Header';
 export { WalletButton } from './WalletButton';
 export { NetworkButton } from './NetworkButton';
+export { TokenPicker } from './TokenPicker';
